refactored message menu
