+++ tsconfig.json
@@ -19,6 +19,6 @@
     "module": "esnext",
     "isolatedModules": true
   },
-  "include": ["src"]
-  // "include": ["src/index.web.tsx","web"]
+  // "include": ["src"]
+  "include": ["src/index.web.tsx","web"]
 }
