+++ web/chat/chat.tsx
@@ -10,16 +10,10 @@ import { useObserver } from 'mobx-react-lite'
 import Frame from './frame'
 import { CircularProgress } from '@material-ui/core'
 import Alert from '@material-ui/lab/Alert';
-import Menu from '@material-ui/core/Menu';
-import MenuItem from '@material-ui/core/MenuItem';
-import ReplyIcon from '@material-ui/icons/Reply';
-import LinkIcon from '@material-ui/icons/Link';
-import DeleteForeverIcon from '@material-ui/icons/DeleteForever';
-import Button from '@material-ui/core/Button';
 import Dropzone from 'react-dropzone'
 import { uploadFile } from '../utils/meme'
 import Bots from './bots'
-import { useHasLink } from './msg/hooks'
+import MsgMenu from './msgMenu'
 const { useMsgs } = hooks
 
 var link = null
@@ -103,7 +97,6 @@ function ChatContent({ appMode, footHeight }) {
   const [anchorEl, setAnchorEl] = React.useState(null);
   const [menuMessage, setMenuMessage] = useState(null)
   const [uploading, setUploading] = useState(false)
-  const [deleting, setDeleting] = useState(false)
   const [msgCount, setMsgCount] = useState(20)
 
   async function dropzoneUpload(files) {
@@ -145,23 +138,13 @@ function ChatContent({ appMode, footHeight }) {
     const chat = ui.selectedChat
     const appURL = ui.applicationURL
 
-    const link = useHasLink(menuMessage)
-
-    async function deleteMessage() {
-      if (deleting) return
-      setDeleting(true)
-      await msg.deleteMessage(menuMessage.id)
-      setDeleting(false)
-      handleMenuClose()
-    }
-
     const msgs = useMsgs(chat) || []
     const isTribe = chat && chat.type === constants.chat_types.tribe
     const h = `calc(100% - ${headHeight + footHeight}px)`
 
-    useEffect(()=> {
+    useEffect(() => {
       setMsgCount(20)
-    }, [(chat && chat.id)]) 
+    }, [(chat && chat.id)])
 
     if (ui.loadingChat) {
       return <LoadingWrap style={{ maxHeight: h, minHeight: h }}>
@@ -212,44 +195,9 @@ function ChatContent({ appMode, footHeight }) {
                   })}
                 </MsgList>
                 {alert && <Alert style={{ position: 'absolute', bottom: 20, left: 'calc(50% - 90px)', opacity: 0.7, height: 35, padding: `0px 8px 4px 8px` }} icon={false}>{alert}</Alert>}
-                <Menu
-                  id="simple-menu" anchorEl={anchorEl} getContentAnchorEl={null} keepMounted open={Boolean(anchorEl)} onClose={handleMenuClose}
-                  anchorOrigin={{
-                    vertical: 'bottom',
-                    horizontal: isMe(menuMessage) ? 'left' : 'right',
-                  }}
-                  transformOrigin={{
-                    vertical: 'top',
-                    horizontal: isMe(menuMessage) ? 'right' : 'left',
-                  }}
-                  PaperProps={{
-                    style: {
-                      backgroundColor: isMe(menuMessage) ? theme.highlight : theme.extraDeep
-                    },
-                  }}>
-                  <MenuItem onClick={() => { navigator.clipboard.writeText(menuMessage.message_content), handleMenuClose(), onCopy('Text') }}
-                    style={{ fontSize: 14, color: 'white', backgroundColor: isMe(menuMessage) ? theme.highlight : theme.extraDeep }}>
-                    <svg xmlns="http://www.w3.org/2000/svg" width="15" height="15" viewBox="0 0 24 24" style={{ fill: 'white', marginRight: 8 }}>
-                      <path d="M16 1H4c-1.1 0-2 .9-2 2v14h2V3h12V1zm3 4H8c-1.1 0-2 .9-2 2v14c0 1.1.9 2 2 2h11c1.1 0 2-.9 2-2V7c0-1.1-.9-2-2-2zm0 16H8V7h11v14z" />
-                    </svg>Copy Text
-                  </MenuItem>
-                    {link &&
-                      <MenuItem onClick={() => { navigator.clipboard.writeText(link), handleMenuClose(), onCopy('Link') }}
-                        style={{ fontSize: 14, color: 'white', backgroundColor: isMe(menuMessage) ? theme.highlight : theme.extraDeep }}>
-                        <LinkIcon style={{ fontSize: 'medium', marginRight: 8 }} />Copy Link
-                  </MenuItem>}
-                  <MenuItem onClick={() => { ui.setReplyUUID(menuMessage.uuid), handleMenuClose() }}
-                    style={{ fontSize: 14, color: 'white', backgroundColor: isMe(menuMessage) ? theme.highlight : theme.extraDeep }}>
-                    <ReplyIcon style={{ fontSize: 'medium', marginRight: 8 }} />Reply
-                  </MenuItem>
-
-                  {isMe(menuMessage) &&
-                    <MenuItem onClick={deleteMessage}
-                      style={{ fontSize: 14, color: '#fe5251', backgroundColor: isMe(menuMessage) ? theme.highlight : theme.extraDeep }}>
-                      <DeleteForeverIcon style={{ color: 'red', fontSize: 'medium', marginRight: 8 }} />
-                      {deleting ? 'Deleting...' : 'Delete Message'}
-                    </MenuItem>}
-                </Menu>
+                <MsgMenu anchorEl={anchorEl} menuMessage={menuMessage} isMe={isMe}
+                  handleMenuClose={handleMenuClose} onCopy={onCopy}
+                />
               </Layer>
               {appURL && <Layer show={appMode} style={{ background: theme.deep, height: 'calc(100% + 63px)' }}>
                 <Frame url={appURL} />
